First Try to Construct a Matrix Class
Matrix Header File with Error. Initially I would like to borrow the idea from the practice final and use a one dimensional array to implement the matrix class. However, when incrementing the iterator of the class, the process can be uselessly hard. Hence I gave up the idea and decide to try a 2-D array to implement the matrix class.